[Update] StarrySky 백그라운드 추가 적용 ( 권한요청 화면, 온보딩 화면 )
